feat(db): implement CRUD operations for assistants, sessions, messages, groups, file stores, contents, and chunks
- Added CRUD functionality for managing assistants, sessions, messages, groups, file stores, file contents, and file chunks.
- Introduced LocalDatabase class using Dexie for database interactions.
- Created utility functions for pagination and bulk operations.
- Established a centralized database service for easier access to CRUD operations.
- Enhanced database schema with versioning and upgrades for better data management.
- Refactored MCP worker to utilize static imports for module loading, improving bundling and type safety.
