+++ docs/history/refactoring_20250831_1600.md
@@ -766,17 +766,216 @@ src/features/chat/
 │   └── ...
 ```
 
-### Phase 4: Rust 파일들 리팩터링 (3일)
+### Phase 4: Rust Backend 리팩터링 (filesystem.rs, mcp.rs) (3일)
 
-1. `filesystem.rs`: schema_builder.rs 헬퍼 활용
-2. `mcp.rs`: 타입/서버/유틸리티 모듈 분리
-3. Rust 빌드 및 테스트 검증
+#### Phase 4 작업의 목적
 
-### Phase 5: db.ts 리팩터링 (2일)
+`filesystem.rs`와 `mcp.rs`의 대형 단일 파일을 기능별로 모듈화하여 코드 중복을 제거하고 유지보수성을 향상시키는 것. JSONSchema 생성, 파일 I/O 보안 검증, MCPTool 등록/관리, 서버 연결/샘플링/도구 호출 등 각 기능을 별도 모듈로 분리하여 재사용성과 테스트 용이성을 강화한다. Rust 코딩 규약 준수 및 타입 안전성을 보장한다.
 
-1. 타입 정의 분리 (`types.ts`)
-2. CRUD 인터페이스 분리 (`crud.ts`)
-3. DB 서비스 분리 (`service.ts`)
+#### Phase 4 현재의 상태 / 문제점
+
+- **단일 책임 위반**: 각 파일이 다중 책임 담당 (타입 정의, 도구 생성, 서버 관리, 유틸리티 등 혼재)
+- **코드 중복**: JSONSchema 생성 패턴 반복, 파일 경로 검증 로직 중복, MCPTool 등록 패턴 반복
+- **유지보수 어려움**: 800줄 이상의 대형 파일로 기능 추가/수정이 어려움
+- **테스트 어려움**: 단일 파일로 인한 단위 테스트 및 Mock 주입 어려움
+- **타입 안전성 부족**: 일부 타입 추론에 의존, 인터페이스 분리 미흡
+
+#### Phase 4 추가 분석 과제
+
+- 모듈 분리 후 의존성 순환 참조 발생 가능성 분석
+- 기존 MCPTool 인터페이스 호환성 유지 여부 검증
+- 파일 I/O 보안 검증 로직의 일관성 유지 방법 분석
+- Rust 빌드 및 컴파일 시간 영향 측정
+
+#### Phase 4 변경 이후의 상태 / 해결 판정 기준
+
+- **빌드/테스트 통과**: `cargo build && cargo test` 성공
+- **모듈 분리 완료**: 각 파일의 책임이 단일하게 분리, 서브모듈 5개 이상 생성
+- **코드 중복 제거**: JSONSchema/파일 검증/MCPTool 생성 코드 70% 이상 감소
+- **타입 안전성 보장**: 모든 파일에서 타입 추론 최소화, 명시적 타입 정의
+- **기능 유지**: 기존 MCP 서버 연결, 도구 호출, 파일 I/O 정상 동작
+
+#### Phase 4 수정이 필요한 코드 및 수정부분의 코드 스니핏
+
+##### 1. filesystem.rs MCPTool 생성 함수 분리
+
+```rust
+// 변경 전: filesystem.rs (현재 - 842줄)
+fn create_read_file_tool() -> MCPTool {
+    MCPTool {
+        input_schema: JSONSchema {
+            schema_type: JSONSchemaType::Object {
+                properties: Some({
+                    let mut props = HashMap::new();
+                    props.insert("path".to_string(), JSONSchema {
+                        schema_type: JSONSchemaType::String {
+                            min_length: Some(1),
+                            max_length: Some(1000),
+                            description: Some("Path to file".to_string()),
+                        },
+                        required: true,
+                    });
+                    // ... 20줄 이상 반복
+                    props
+                }),
+                required: Some(vec!["path".to_string()]),
+            },
+        },
+        // ...
+    }
+}
+```
+
+```rust
+// 변경 후: filesystem.rs (리팩토링 후)
+use crate::mcp::utils::schema_builder::{string_prop, integer_prop, object_schema};
+
+fn create_read_file_tool() -> MCPTool {
+    let mut props = HashMap::new();
+    props.insert("path".to_string(), string_prop(Some(1), Some(1000), Some("Path to file")));
+    props.insert("start_line".to_string(), integer_prop(Some(1), None, Some("Start line")));
+
+    MCPTool {
+        input_schema: object_schema(props, vec!["path".to_string()]),
+        // 간결화된 코드 (기존 30줄 → 5줄)
+    }
+}
+```
+
+##### 2. mcp.rs 타입/서버/유틸리티 분리
+
+```rust
+// 변경 전: mcp.rs (현재 - 834줄)
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct MCPServerConfig { ... }
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub enum JSONSchemaType { ... }
+
+// ... 서버 관리 함수들
+// ... 유틸리티 함수들
+```
+
+```rust
+// 변경 후: 분리된 파일들
+// types.rs
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct MCPServerConfig { ... }
+
+// schema.rs
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub enum JSONSchemaType { ... }
+
+// server.rs
+pub struct MCPServerManager {
+    // 서버 관리 로직만
+}
+
+// utils.rs
+// 유틸리티 함수들만
+```
+
+#### Phase 4 재사용 가능한 연관 코드
+
+##### 1. schema_builder.rs (JSONSchema 생성 헬퍼)
+
+**파일 경로**: `src-tauri/src/mcp/utils/schema_builder.rs`
+
+**주요 기능**:
+
+- `string_prop(min_len, max_len, description)`: 문자열 속성 생성
+- `integer_prop(min, max, description)`: 정수 속성 생성
+- `object_schema(properties, required)`: 객체 스키마 생성
+- `array_schema(items, description)`: 배열 스키마 생성
+
+**특징**:
+
+- JSONSchema 생성 코드 중복 제거
+- 타입 안전성 보장
+- 재사용성 높음
+
+##### 2. secure_file_manager.rs (파일 I/O 및 보안)
+
+**파일 경로**: `src-tauri/src/services/secure_file_manager.rs`
+
+**주요 기능**:
+
+- `validate_path(path)`: 파일 경로 보안 검증
+- `write_file_string(path, content)`: 안전한 파일 쓰기
+- `read_file_string(path)`: 안전한 파일 읽기
+- `SecurityValidator`: 보안 검증 인터페이스
+
+**특징**:
+
+- 파일 시스템 보안 보장
+- 크로스 플랫폼 호환성
+- 에러 처리 표준화
+
+##### 3. utils.rs (상수 및 보안 검증)
+
+**파일 경로**: `src-tauri/src/mcp/builtin/utils.rs`
+
+**주요 기능**:
+
+- `MAX_FILE_SIZE`: 최대 파일 크기 상수
+- `SecurityValidator`: 보안 검증 구현
+- 파일 타입 검증 헬퍼
+
+**특징**:
+
+- 중앙화된 상수 관리
+- 재사용 가능한 보안 검증 로직
+
+##### 4. MCPServerManager 및 관련 타입
+
+**파일 경로**: `src-tauri/src/mcp.rs`
+
+**주요 기능**:
+
+- `MCPServerManager`: 서버 연결/관리
+- `MCPTool`, `MCPResponse`: MCP 프로토콜 타입
+- `start_server()`, `call_tool()`: 서버 조작 함수
+
+**특징**:
+
+- MCP 프로토콜 표준 준수
+- 비동기 처리 지원
+- 에러 처리 체계화
+
+#### Phase 4 변경 후 파일 구조
+
+```bash
+src-tauri/src/mcp/
+├── types.rs                 # MCPServerConfig, MCPTool 등 타입 정의
+├── schema.rs                # JSONSchemaType, JSONSchema 등 스키마 타입
+├── server.rs                # MCPServerManager, 서버 관리 로직
+├── utils.rs                 # 유틸리티 함수들
+└── builtin/
+    ├── filesystem.rs        # 간결화된 파일 시스템 도구 (schema_builder 활용)
+    └── mod.rs
+
+src-tauri/src/services/
+└── secure_file_manager.rs   # 파일 I/O 보안 서비스
+
+src-tauri/src/mcp/utils/
+└── schema_builder.rs        # JSONSchema 생성 헬퍼
+```
+
+### Phase 5: db.ts 리팩터링 (2일) ✅ **COMPLETED**
+
+**작업 완료 내용:**
+1. **타입 정의 분리** (`src/lib/db/types.ts`) - 모든 인터페이스 및 타입 정의를 별도 파일로 분리
+2. **CRUD 인터페이스 분리** (`src/lib/db/crud.ts`) - 각 테이블별 CRUD 구현을 모듈화
+3. **DB 서비스 분리** (`src/lib/db/service.ts`) - LocalDatabase 클래스와 dbUtils 유틸리티 분리
+4. **인덱스 파일 생성** (`src/lib/db/index.ts`) - 통합 export 제공
+5. **기존 호환성 유지** (`src/lib/db.ts`) - 기존 import 경로 유지를 위한 re-export
+
+**결과:**
+- 기존 841줄의 단일 파일을 4개 모듈로 분리 (평균 200줄 내외)
+- 타입 정의, CRUD 로직, DB 서비스가 각각 단일 책임으로 분리됨
+- 모든 기존 import (`from '@/lib/db'`)가 변경 없이 동작
+- `pnpm lint && pnpm format && pnpm build` 모두 통과
+- 코드 중복 제거 및 유지보수성 향상
 
 ### Phase 6: 통합 테스트 및 검증 (2일)
