+++ src/lib/db/index.ts
@@ -1,0 +1,23 @@
+export type {
+  CRUD,
+  DatabaseObject,
+  DatabaseService,
+  FileChunk,
+  FileContent,
+  FileContentCRUD,
+  FileStore,
+  Page,
+} from './types';
+
+export { dbService, dbUtils, LocalDatabase } from './service';
+
+export {
+  assistantsCRUD,
+  fileChunksCRUD,
+  fileContentsCRUD,
+  fileStoresCRUD,
+  groupsCRUD,
+  messagesCRUD,
+  objectsCRUD,
+  sessionsCRUD,
+} from './crud';
